sm: Drop switching PF5302 operation mode to OFF
This is a workaround to fix the following issue:
Issuing the reset command from U-Boot prompt the DART-MX94 reset successfully the first
13 times, the 14th time the reset doesn't work, at this point:
- U-Boot shell is not responsive
- The reset KEY (SW5 DT8MCB) does not work.
- The SM shell is not responsive

The reason of this behavior is that during the reset command the switching mode of the
DC-DC converter inside the PF5302 is changed from OFF to PWM. The I2C register that
controls the PF5302 operating mode is SW1_CTRL1 (address 0x0F, bit3, bit2).
SW1 Mode Selection:
bit:  3    2
      0    0  Off
      0    1  Reserved
      1    0  Reserved
      1    1  PWM

The NXP commit which introduced the use of operation mode "OFF" for PF5302 is:
 - 2d5c1c1fb

To fix the issue this patch avoid to switch the PF5302 operating mode to OFF during
reset command.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>

Conflicts:
	sm/lmm/lmm_sys.c

(cherry picked from commit a0a907ee83c842ae78cc87babbee34a592ea331b)
Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>